use more restrictive permissions for chrome

+++ src/utils/api/cloud-api.ts
@@ -85,6 +85,15 @@ const getUrl = (options: Options, url: URLS, query?: URLSearchParams) => {
     return `${baseUrl}${pathInfo.url}${queryString}`
 }
 
+export const getDomains = (options: Partial<Options>) => [`${options.domain}/*`, 'https://*.tempo.io/*']
+export const checkPermissions = async (options: Partial<Options>) => {
+    if (isFirefox) return true
+    const permissions = (isFirefox ? browser : chrome)?.permissions
+    if (!permissions) return Promise.reject('Unable to access permission api.')
+
+    return new Promise((resolve, reject) => permissions.contains({ origins: getDomains(options) }, (hasPermission) => hasPermission ? resolve(true) : reject('No permission to access the api.')))
+}
+
 const createWorklogPayload = (options: Options, worklog: Partial<TemporaryWorklog | Worklog>): WorklogPayload => {
     const { end, start, issue, id, comment } = worklog
     const seconds = Math.round((end - start) / 1000)
@@ -145,6 +154,7 @@ function toLocalWorklog(options: Options, simpleMapping?: boolean) {
 }
 
 export async function fetchIssueById(options, id: number): Promise<RemoteIssue> {
+    await checkPermissions(options)
     const query = new URLSearchParams()
     query.append('fields', 'summary')
     const url = getUrl(options, URLS.ISSUE)
@@ -152,6 +162,7 @@ export async function fetchIssueById(options, id: number): Promise<RemoteIssue> 
 }
 
 export async function fetchIssues(options, jql, limit?: number): Promise<Issue[]> {
+    await checkPermissions(options)
     const query = new URLSearchParams()
     query.append('jql', jql)
     if (limit) {
@@ -164,8 +175,9 @@ export async function fetchIssues(options, jql, limit?: number): Promise<Issue[]
 }
 
 export async function fetchSelf(options?: Options, useCredentials?: boolean): Promise<Self> {
+    await checkPermissions(options)
     const result = await fetch(getUrl(options, URLS.SELF), {
-        headers:  useCredentials ? jsonHeaders : headers(options, URLS.SELF),
+        headers: useCredentials ? jsonHeaders : headers(options, URLS.SELF),
         credentials: useCredentials ? 'include' : 'omit',
         redirect: 'manual'
     })
@@ -177,6 +189,7 @@ export async function fetchSelf(options?: Options, useCredentials?: boolean): Pr
 }
 
 export async function searchIssues(options: Options, searchString: string): Promise<string[]> {
+    await checkPermissions(options)
     const query = new URLSearchParams()
     query.append('query', searchString)
     query.append('currentJQL', 'created >= "1900/01/01"')
@@ -195,6 +208,7 @@ export async function fetchWorklogs(
     options: Options,
     simpleMapping?: boolean
 ): Promise<Worklog[]> {
+    await checkPermissions(options)
     const payload = {
         from: dateString(startDate),
         to: dateString(endDate),
@@ -220,6 +234,7 @@ export async function fetchWorklogs(
 }
 
 export async function writeWorklog(worklog: Partial<Worklog>, options?: Options): Promise<Worklog> {
+    await checkPermissions(options)
     const payload = createWorklogPayload(options, worklog)
 
     return fetchJson(getUrl(options, URLS.CREATE_WORKLOG), {
@@ -231,6 +246,7 @@ export async function writeWorklog(worklog: Partial<Worklog>, options?: Options)
 }
 
 export async function updateWorklog(worklog: Partial<Worklog>, options?: Options): Promise<Worklog> {
+    await checkPermissions(options)
     const payload = createWorklogPayload(options, worklog)
     return fetchJson(`${getUrl(options, URLS.CREATE_WORKLOG)}/${payload.tempoWorklogId}`, {
         headers: headers(options, URLS.CREATE_WORKLOG),
@@ -241,6 +257,7 @@ export async function updateWorklog(worklog: Partial<Worklog>, options?: Options
 }
 
 export async function deleteWorklog(id: string, options: Options): Promise<void> {
+    await checkPermissions(options)
     return fetch(`${getUrl(options, URLS.CREATE_WORKLOG)}/${id}`, {
         headers: headers(options, URLS.CREATE_WORKLOG),
         method: 'DELETE',
